requirements.txt
# LangChain 核心依賴
langchain>=0.1.14
langchain-core>=0.1.33
langchain-community>=0.0.38

# LangChain 整合
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.1
langchain-qdrant>=0.1.0

# 向量資料庫
qdrant-client>=2.7.3

# Embedding 模型 - Mac M4 Pro 優化版本
sentence-transformers>=2.2.0

# PyTorch for Apple Silicon (M4 Pro) - 支援 MPS 加速
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Apple Silicon 加速支援
accelerate>=0.20.0

# LLM 服務
ollama>=0.1.25

# Web UI
gradio>=4.26.0

# 配置管理
python-dotenv>=1.0.0
pyyaml>=6.0.1

# 工具類
requests>=2.31.0
numpy>=1.24.0

# 測試依賴（可選）
pytest>=7.4.3
pytest-asyncio>=0.23.2
hypothesis>=6.92.1
